requirements.txt
webdriver-manager==4.0.2
selenium==4.37.0
python-dotenv
pytest
playwright==1.55.0
faker==37.11.0
pytest-playwright-visual-snapshot==0.2.3
pytest-repeat==0.9.4
pytest-cases==3.9.1
pytest-variables==3.1.0
pytest-html==4.1.1
pytest-xdist==3.8.0
pytest-timeout==2.4.0
pytest-metadata==3.1.1
pytest-randomly==4.0.1
pytest-dependency==0.6.0
pytest-dotenv==0.5.2
pytest-sugar==1.1.1
pytest-base-url==2.1.0
pytest-bdd==8.1.0
pytest-rerunfailures==16.1
pytest-selenium==4.1.0
requests==2.32.5
pytest-clarity==1.0.1
pytest_check==2.6.0

fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.1.1
pyjwt==2.8.0
python-multipart==0.0.19
email-validator==2.1.0

# Testing
pytest-asyncio==0.21.1
httpx==0.25.2
allure-pytest==2.15.3
Pillow>=11.1.0

# dev
black== 25.9.0
flake8==7.3.0
isort==7.0.0
pre-commit==4.3.0
